enh: Switch to custom heap implementation
Since we know a query has k max particles, instead of using the list-based
heapq implementation, we can use an array-based implementation that doesn't
rely on a list of tuples. It's somewhat simpler because we can use the
max-heap version (unsupported in numba) and we only need to implement the
heapreplace method and have a way of getting sorted results.
Of course, the actual implementation is slightly more complicated, since we
need to ensure every swap is true of both the distances and indices arrays,
instead of just the distance-index tuple.
But with these changes, our query implementation is only 10x slower than scipy
instead of 100x slower :/

+++ packingcubes/packed_tree/packed_tree.py
@@ -412,7 +412,7 @@ class PackedTree(octree.Octree):
         k: int = 1,
         use_data_indices: bool = True,
         return_sorted: bool = True,
-    ) -> tuple[NDArray[np.float64], NDArray[np.uint32]]:
+    ) -> tuple[NDArray[np.float64], NDArray[np.int_]]:
         """
         Get kth nearest particle distances and indices to point
